Modify Dto fileName (LottoGetReponse) Modify member type In LottoHistoryEntity round ( long -> int ) Add get LottoHistoryEntity Iin Repository( by round, by drawDate)

+++ src/main/java/com/lotto/web/model/entity/LottoHistoryEntity.java
@@ -21,7 +21,7 @@ public class LottoHistoryEntity extends CreationTimestampEntity{
     private int fourthNumber;
     private int fifthNumber;
     private int sixthNumber;
-    private long round;
+    private int round;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date drawDate;
